+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Populate database with genes downloaded from Ensembl Biomart
 - Function to create a VCF Bedtool filter from a list of genes HGNC ids or Ensembl ids
 - API endpoint to add variants using a POST request
+- API endpoint to remove variants using POST request
 
 
 ## [1.1] - 2020.06.17